+++ src/bin/worker.ts
@@ -8,14 +8,26 @@ import { CONFETTI_PID_PATH, DEFAULT_PORT } from '../util/constants';
 import { debug } from '../logger/logger';
 import loadConfigurationFile from '../util/loadConfigurationFile';
 
-const config = process.env.CONFETTI_CONFIG_PATH ? loadConfigurationFile(process.env.CONFETTI_CONFIG_PATH) : getGlobalConfig();
+const config = process.env.CONFETTI_CONFIG_PATH
+    ? loadConfigurationFile(process.env.CONFETTI_CONFIG_PATH)
+    : getGlobalConfig();
 
-fs.writeFileSync(CONFETTI_PID_PATH, process.pid.toString(10))
+if (fs.existsSync(CONFETTI_PID_PATH)) {
+    throw new Error('Worker already running');
+}
 
-process.on("exit", () => {
-    fs.unlinkSync(CONFETTI_PID_PATH);
+fs.writeFileSync(CONFETTI_PID_PATH, process.pid.toString(10));
+
+function onExit() {
+    if (fs.existsSync(CONFETTI_PID_PATH)) {
+        fs.unlinkSync(CONFETTI_PID_PATH);
+    }
     process.exit();
-});
+}
+
+process.on('exit', onExit);
+process.on('SIGINT', onExit);
+process.on('SIGTERM', onExit);
 
 const port = config.port || DEFAULT_PORT;
 const app = createApp(config);
